+++ modules/pool/pool-debug.test.ts
@@ -376,4 +376,72 @@ describe('pool debugging', () => {
         // await poolService.syncAllPoolsFromSubgraph();
         // await poolService.syncChangedPools();
     }, 5000000);
+
+    it('pool hook data', async () => {
+        initRequestScopedContext();
+        setRequestScopedContextValue('chainId', '11155111');
+
+        const pool = await prisma.prismaPool.findFirst({
+            where: { id: '0x75f49d54978d08e4e76a873da6c78e8f6b2901c2', chain: 'SEPOLIA' },
+        });
+
+        const gqlPool = await poolService.getGqlPool('0x75f49d54978d08e4e76a873da6c78e8f6b2901c2', 'SEPOLIA');
+
+        console.log(gqlPool.id);
+
+        //only do once before starting to debug
+        // await poolService.syncAllPoolsFromSubgraph();
+        // await poolService.syncChangedPools();
+    }, 5000000);
+
+    it('pool tokens minimalpool', async () => {
+        initRequestScopedContext();
+        setRequestScopedContextValue('chainId', '11155111');
+
+        const gqlPools = await poolService.getGqlPools({where: {
+            "poolTypeIn": [
+                "WEIGHTED",
+                "STABLE",
+                "COMPOSABLE_STABLE",
+                "META_STABLE",
+                "LIQUIDITY_BOOTSTRAPPING",
+                "GYRO",
+                "GYRO3",
+                "GYROE",
+                "COW_AMM",
+                "FX"
+            ],
+            "chainIn": [
+                "MAINNET",
+                "ARBITRUM",
+                "AVALANCHE",
+                "BASE",
+                "GNOSIS",
+                "POLYGON",
+                "ZKEVM",
+                "OPTIMISM",
+                "MODE",
+                "FRAXTAL",
+                "SEPOLIA"
+            ],
+            "userAddress": null,
+            "minTvl": 0,
+            "tagIn": null,
+            "tagNotIn": [
+                "BLACK_LISTED"
+            ],
+            "protocolVersionIn": [
+                3
+            ]
+        },
+        "textSearch": null
+    })
+    
+
+        console.log(gqlPools.length);
+
+        //only do once before starting to debug
+        // await poolService.syncAllPoolsFromSubgraph();
+        // await poolService.syncChangedPools();
+    }, 5000000);
 });