+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # backend
 
+## 1.23.1
+
+### Patch Changes
+
+-   7f0ee69: adapt nested balance and usd of a token to its share
+-   6d14f60: enable LBP in sor and sync
+
+## 1.23.0
+
+### Minor Changes
+
+-   23f6773: Add hook and erc4626 reviews and tags. Adapt to new metadata repo layout.
+
+### Patch Changes
+
+-   3c0909d: use new priceimpact calc for Beets SOR path
+-   5cabb48: sync bptPriceRate for v3 pools
+-   33ffa59: reduce vebal balance multicall batchsize
+-   daa73e2: fix review data exposure
+-   4337d11: return all token prices
+-   1b3ef2d: fix sor hops with buffers
+-   232ee9e: add tokenIn filter for tokens query
+
 ## 1.22.0
 
 ### Minor Changes